Wip more tests
