+++ packages/perspective-viewer-d3fc/test/js/area.spec.js
@@ -10,8 +10,7 @@
 const path = require("path");
 
 // const utils = require("@finos/perspective-test");
-const simple_tests = require("@finos/perspective-viewer/test/js/simple_tests.js");
-const utils = require("@finos/perspective-test");
+const testutils = require("@finos/perspective-test");
 
 import { test, expect } from "@playwright/test";
 
@@ -19,18 +18,16 @@ const { withTemplate } = require("./simple-template");
 withTemplate("area", "Y Area");
 
 test.describe("Area Tests", () => {
-    simple_tests.easy("area.html", (page) =>
-        utils.get_contents(
+    testutils.easy("area.html", (page) =>
+        testutils.get_contents(
             page,
             "perspective-viewer perspective-viewer-d3fc-yarea"
         )
     );
 
     test("Can click a button in the Shadow DOM", async ({ page }, testInfo) => {
-        await simple_tests.setupPage(page, testInfo, "area.html");
+        await testutils.setupPage(page, testInfo, "area.html");
         let buttonInShadowDom = await page.waitForSelector("#settings_button");
         await buttonInShadowDom.click();
-        return true;
-        // return await get_contents(page);
     });
 });
