+++ tools/perspective-test/src/js/paths.js
@@ -13,6 +13,5 @@ const path = require("path");
 module.exports.IS_LOCAL_PUPPETEER = fs.existsSync(
     path.join(__dirname, "..", "..", "..", "..", "node_modules", "puppeteer")
 );
-module.exports.RESULTS_TAGNAME = `results`;
-module.exports.RESULTS_FILENAME = `${module.exports.RESULTS_TAGNAME}.json`;
-module.exports.RESULTS_DEBUG_FILENAME = `${module.exports.RESULTS_TAGNAME}.debug.json`;
+module.exports.RESULTS_FILENAME = `results.json`;
+module.exports.RESULTS_DEBUG_FILENAME = `results.debug.json`;
